migrated database from version 1 to version 2.

+++ app/src/main/java/com/github/wnebyte/workoutapp/ui/exercisecreate/ExerciseCreateFragment.kt
@@ -122,7 +122,7 @@ class ExerciseCreateFragment: Fragment() {
         }
         binding.timer.doOnTextChanged { text, _, _, count ->
             if ((text != null) && (0 < count)) {
-                exercise.exercise.timer = text.toString().toInt()
+                exercise.exercise.timer = text.toString().toLong()
             }
         }
     }
