+++ app/src/main/java/com/github/wnebyte/workoutapp/ui/exercisedetails/ExerciseDetailsFragment.kt
@@ -121,7 +121,7 @@ class ExerciseDetailsFragment: Fragment() {
         }
         binding.timer.doOnTextChanged { text, _, _, count ->
             if ((text != null) && (0 < count)) {
-                exercise.exercise.timer = text.toString().toInt()
+                exercise.exercise.timer = text.toString().toLong()
             }
         }
     }
